import of CINT 5.15.101. For what else is new see http://root.cern.ch/root/Cint.phtml?relnote.
git-svn-id: http://root.cern.ch/svn/root/trunk@7235 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/include/stddef.h
@@ -6,5 +6,5 @@ typedef int ptrdiff_t;
 typedef long ptrdiff_t;
 #endif
 typedef unsigned int size_t;
-typedef unsigned int wchar_t;
+typedef unsigned short wchar_t;
 #endif
